apps/falcon/requirements.txt
# Core Trading Libraries
backtrader==1.9.78.123
pandas==2.0.3
numpy==1.24.3

# Data Feed (Massive.com / Polygon.io)
requests>=2.32.3

# Web Dashboard Backend
Flask>=3.1.2
Flask-CORS>=4.0.1

# Data Analysis
matplotlib==3.7.2
yfinance==0.2.28

# AI Stock Screener
beautifulsoup4==4.12.2
schedule==1.2.0
pytz==2023.3
python-dotenv==1.0.0
PyYAML>=6.0  # Multi-screener profile YAML export/import

# HTML Parsing
lxml==4.9.3

# Technical Analysis (Optional)
# TA-Lib==0.4.28

# Database
# sqlite3 comes with Python standard library
# PostgreSQL (optional - only install if using PostgreSQL)
# psycopg2-binary==2.9.9

# Development Tools (Optional)
# jupyter==1.0.0
# notebook==7.0.6

# YouTube Strategy Extraction
youtube-transcript-api==1.2.3
yt-dlp>=2025.12.8
anthropic>=0.75.0
